requirements.txt
pytest
selenium
mysql-connector
